+++ web/src/js/__tests__/components/Header/MenuToggleSpec.tsx
@@ -5,11 +5,7 @@ import {
     MenuToggle,
     OptionsToggle,
 } from "../../../components/Header/MenuToggle";
-import { Provider } from "react-redux";
-import { TStore } from "../../ducks/tutils";
-import * as optionsEditorActions from "../../../ducks/ui/optionsEditor";
 import { fireEvent, render, screen, waitFor } from "../../test-utils";
-import { fetchApi, runCommand } from "../../../utils";
 
 import { enableFetchMocks } from "jest-fetch-mock";
 
@@ -17,13 +13,13 @@ enableFetchMocks();
 
 describe("MenuToggle Component", () => {
     it("should render correctly", () => {
-        let changeFn = jest.fn(),
-            menuToggle = renderer.create(
-                <MenuToggle onChange={changeFn} value={true}>
-                    <p>foo children</p>
-                </MenuToggle>,
-            ),
-            tree = menuToggle.toJSON();
+        const changeFn = jest.fn();
+        const menuToggle = renderer.create(
+            <MenuToggle onChange={changeFn} value={true}>
+                <p>foo children</p>
+            </MenuToggle>,
+        );
+        const tree = menuToggle.toJSON();
         expect(tree).toMatchSnapshot();
     });
 });