+++ web/src/js/__tests__/components/Modal/OptionModalSpec.tsx
@@ -4,18 +4,18 @@ import { PureOptionDefault } from "../../../components/Modal/OptionModal";
 
 describe("PureOptionDefault Component", () => {
     it("should return null when the value is default", () => {
-        let pureOptionDefault = renderer.create(
-                <PureOptionDefault value="foo" defaultVal="foo" />,
-            ),
-            tree = pureOptionDefault.toJSON();
+        const pureOptionDefault = renderer.create(
+            <PureOptionDefault value="foo" defaultVal="foo" />,
+        );
+        const tree = pureOptionDefault.toJSON();
         expect(tree).toMatchSnapshot();
     });
 
     it("should handle boolean type", () => {
-        let pureOptionDefault = renderer.create(
-                <PureOptionDefault value={true} defaultVal={false} />,
-            ),
-            tree = pureOptionDefault.toJSON();
+        const pureOptionDefault = renderer.create(
+            <PureOptionDefault value={true} defaultVal={false} />,
+        );
+        const tree = pureOptionDefault.toJSON();
         expect(tree).toMatchSnapshot();
     });
 
@@ -37,18 +37,18 @@ describe("PureOptionDefault Component", () => {
     });
 
     it("should handle string", () => {
-        let pureOptionDefault = renderer.create(
-                <PureOptionDefault value="foo" defaultVal="" />,
-            ),
-            tree = pureOptionDefault.toJSON();
+        const pureOptionDefault = renderer.create(
+            <PureOptionDefault value="foo" defaultVal="" />,
+        );
+        const tree = pureOptionDefault.toJSON();
         expect(tree).toMatchSnapshot();
     });
 
     it("should handle null value", () => {
-        let pureOptionDefault = renderer.create(
-                <PureOptionDefault value="foo" defaultVal={null} />,
-            ),
-            tree = pureOptionDefault.toJSON();
+        const pureOptionDefault = renderer.create(
+            <PureOptionDefault value="foo" defaultVal={null} />,
+        );
+        const tree = pureOptionDefault.toJSON();
         expect(tree).toMatchSnapshot();
     });
 });