+++ web/src/js/components/EventLog.tsx
@@ -1,22 +1,32 @@
 import React, { Component } from "react";
-import PropTypes from "prop-types";
 import { connect } from "react-redux";
-import { toggleFilter, toggleVisibility } from "../ducks/eventLog";
+import {
+    EventLogItem,
+    LogLevel,
+    toggleFilter,
+    toggleVisibility,
+} from "../ducks/eventLog";
 import ToggleButton from "./common/ToggleButton";
 import EventList from "./EventLog/EventList";
+import { RootState } from "../ducks";
 
-export class PureEventLog extends Component {
-    static propTypes = {
-        filters: PropTypes.object.isRequired,
-        events: PropTypes.array.isRequired,
-        toggleFilter: PropTypes.func.isRequired,
-        close: PropTypes.func.isRequired,
-        defaultHeight: PropTypes.number,
-    };
+type EventLogState = {
+    height: number;
+};
 
+type EventLogProps = {
+    events: EventLogItem[];
+    filters: { [level in LogLevel]: boolean };
+    toggleFilter: (filter: LogLevel) => any;
+    close: () => any;
+    defaultHeight: number;
+};
+
+export class PureEventLog extends Component<EventLogProps, EventLogState> {
     static defaultProps = {
         defaultHeight: 200,
     };
+    private dragStart: number;
 
     constructor(props, context) {
         super(props, context);
@@ -28,7 +38,7 @@ export class PureEventLog extends Component {
         this.onDragStop = this.onDragStop.bind(this);
     }
 
-    onDragStart(event) {
+    onDragStart(event: React.MouseEvent) {
         event.preventDefault();
         this.dragStart = this.state.height + event.pageY;
         window.addEventListener("mousemove", this.onDragMove);
@@ -36,12 +46,12 @@ export class PureEventLog extends Component {
         window.addEventListener("dragend", this.onDragStop);
     }
 
-    onDragMove(event) {
+    onDragMove(event: MouseEvent) {
         event.preventDefault();
         this.setState({ height: this.dragStart - event.pageY });
     }
 
-    onDragStop(event) {
+    onDragStop(event: MouseEvent | DragEvent) {
         event.preventDefault();
         window.removeEventListener("mousemove", this.onDragMove);
     }
@@ -55,16 +65,14 @@ export class PureEventLog extends Component {
                 <div onMouseDown={this.onDragStart}>
                     Eventlog
                     <div className="pull-right">
-                        {["debug", "info", "web", "warn", "error"].map(
-                            (type) => (
-                                <ToggleButton
-                                    key={type}
-                                    text={type}
-                                    checked={filters[type]}
-                                    onToggle={() => toggleFilter(type)}
-                                />
-                            ),
-                        )}
+                        {Object.values(LogLevel).map((type) => (
+                            <ToggleButton
+                                key={type}
+                                text={type}
+                                checked={filters[type]}
+                                onToggle={() => toggleFilter(type)}
+                            />
+                        ))}
                         <i onClick={close} className="fa fa-close"></i>
                     </div>
                 </div>
@@ -75,7 +83,7 @@ export class PureEventLog extends Component {
 }
 
 export default connect(
-    (state) => ({
+    (state: RootState) => ({
         filters: state.eventLog.filters,
         events: state.eventLog.view,
     }),