+++ web/src/js/components/helpers/VirtualScroll.tsx
@@ -24,11 +24,10 @@ export function calcVScroll(
         opts;
     const viewportBottom = viewportTop + viewportHeight;
 
-    let start = 0;
-    let end = 0;
-
-    let paddingTop = 0;
-    let paddingBottom = 0;
+    let start = 0,
+        end = 0,
+        paddingTop = 0,
+        paddingBottom = 0;
 
     if (itemHeights) {
         let pos = 0;
@@ -62,7 +61,7 @@ export function calcVScroll(
     } else {
         // We may have removed a lot of rows since the last render,
         // which means viewportTop will move up.
-        let newViewportTop = Math.min(
+        const newViewportTop = Math.min(
             viewportTop,
             Math.max(0, itemCount * rowHeight - viewportHeight),
         );